chapter_001/src/main/java/ru/job4j/array/Turn.java
package ru.job4j.array;
/**
 * Перевернуть массив.
 *
 * @author Dmitry Rozhdestvenskiy (mailto:dmr1433@gmail.com)
 * @version $Id$
 * @since 0.1
 */
public class Turn {
    /**
     * @param array массив, который следует перевернуть.
     * @return перевёрнутый массив.
     */
    public int[] turn(int[] array) {
        for (int i = 0; i < array.length / 2; i++) {
            int temp = array[i];
            array[i] = array[array.length - i - 1];
            array[array.length - i - 1] = temp;
        }
        return array;
    }
}
